Mapping drone cutscene 1 done.
