+++ Assets/Cinematics/Mapping Drone.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f8d6fbf2cd5213408e4accc43563b4d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
