+++ Assets/Cinematics/Mapping Drone/Mapping Drone Cutscene 1 - Timeline.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fe3c9a624571894f9366e0282ee770f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
