+++ Assets/Cinematics/Signals/FlipPlayer.signal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa786e6786ff7d6418cb947cb4167147
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
