+++ Assets/Cinematics/Signals/UnflipPlayer.signal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: baa4425a04b236b4da14893d403a67d9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
