--- Assets/Scripts/AudioSourceMute.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b8f14aae91be0e4ca4f800d63f22121
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
