--- Assets/Scripts/Camera/CutsceneTrigger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c5627ca07fa2dbb42b8dc7bfac2b52df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
